Refactor memory imports to use advanced memory system module
