+++ core/src/aura_intelligence/inference/free_energy_core.py
@@ -21,7 +21,7 @@ try:
     from ..tda.persistence import TDAProcessor
 except ImportError:
     from ..tda.persistence_simple import TDAProcessor
-from ..memory.hierarchical_memory import HierarchicalMemoryManager
+from ..memory.advanced_memory_system import HierarchicalMemorySystem as HierarchicalMemoryManager
 from ..components.registry import get_registry
 
 import logging
